+++ CHANGES.md
@@ -9,6 +9,7 @@ New Features:
 
 Fixed Issues:
 
+* [#11064](http://dev.ckeditor.com/ticket/11064): [Blink, WebKit] Fixed: Cannot select all when widget or non-editable element is the first/last element of the content. Also fixes this issue in [Select All](http://ckeditor.com/addon/selectall) plugin.
 * [#14755](http://dev.ckeditor.com/ticket/14755): [Blink, WebKit, IE8] Fixed: Browser hangs when table is inserted in place of selected list with empty last item.
 * [#16624](http://dev.ckeditor.com/ticket/16624): Fixed: Background color stripped while [pasting text from Word](http://ckeditor.com/addon/pastefromword). Introduced integration with [Color Button](http://ckeditor.com/addon/colorbutton) plugin which will normalize CSS `background` property if it contains color values when [pasted from Word](http://ckeditor.com/addon/pastefromword).
 * [#16600](http://dev.ckeditor.com/ticket/16600): [Blink, WebKit] Fixed: Error thrown occasionally by uninitialized editable for multiple CKEditor instances on the same page.